+++ keras/backend/numpy_backend/linear_algebra.py
@@ -1,0 +1,66 @@
+import numpy as np
+
+
+def dot(x, y):
+    return np.dot(x, y)
+
+
+def batch_dot(x, y, axes=None):
+    if x.ndim < 2 or y.ndim < 2:
+        raise ValueError('Batch dot requires inputs of rank 2 or more.')
+
+    if isinstance(axes, int):
+        axes = [axes, axes]
+    elif isinstance(axes, tuple):
+        axes = list(axes)
+
+    if axes is None:
+        if y.ndim == 2:
+            axes = [x.ndim - 1, y.ndim - 1]
+        else:
+            axes = [x.ndim - 1, y.ndim - 2]
+
+    if any([isinstance(a, (list, tuple)) for a in axes]):
+        raise ValueError('Multiple target dimensions are not supported. ' +
+                         'Expected: None, int, (int, int), ' +
+                         'Provided: ' + str(axes))
+
+    # Handle negative axes
+    if axes[0] < 0:
+        axes[0] += x.ndim
+    if axes[1] < 0:
+        axes[1] += y.ndim
+
+    if 0 in axes:
+        raise ValueError('Can not perform batch dot over axis 0.')
+
+    if x.shape[0] != y.shape[0]:
+        raise ValueError('Can not perform batch dot on inputs'
+                         ' with different batch sizes.')
+
+    d1 = x.shape[axes[0]]
+    d2 = y.shape[axes[1]]
+    if d1 != d2:
+        raise ValueError('Can not do batch_dot on inputs with shapes ' +
+                         str(x.shape) + ' and ' + str(y.shape) +
+                         ' with axes=' + str(axes) + '. x.shape[%d] != '
+                         'y.shape[%d] (%d != %d).' % (axes[0], axes[1], d1, d2))
+
+    result = []
+    axes = [axes[0] - 1, axes[1] - 1]  # ignore batch dimension
+    for xi, yi in zip(x, y):
+        result.append(np.tensordot(xi, yi, axes))
+    result = np.array(result)
+
+    if result.ndim == 1:
+        result = np.expand_dims(result, -1)
+
+    return result
+
+
+def transpose(x):
+    return np.transpose(x)
+
+
+def gather(reference, indices):
+    return reference[indices]
