+++ keras/backend/numpy_backend/value_manipulation.py
@@ -1,0 +1,7 @@
+def get_value(x):
+    return x
+
+
+def print_tensor(x, message=''):
+    print(x, message)
+    return x
